+++ pkg/model/mq/msg.go
@@ -5,7 +5,7 @@ const PushMsgKey = "easychat_msg_push_topic"
 
 // BroadcastMsg 推送消息体
 type BroadcastMsg struct {
-	Type       int    `json:"type"`       // 消息类型: 1-单聊, 2-群聊
+	Type       int64  `json:"type"`       // 消息类型: 1-单聊, 2-群聊
 	FromUserId int64  `json:"fromUserId"` // 发送人ID
 	ToUserId   int64  `json:"toUserId"`   // 接收人ID（单聊时使用）
 	GroupId    int64  `json:"groupId"`    // 群ID（群聊时使用）← 新增
